+++ packages/commercetools/composables/src/useCart/index.ts
@@ -1,7 +1,6 @@
-import { CustomQueryFn } from '@vue-storefront/commercetools-api';
 import { ProductVariant, Cart, LineItem } from './../types/GraphQL';
 import loadCurrentCart from './currentCart';
-import { AgnosticCoupon, useCartFactory, UseCartFactoryParams, Context } from '@vue-storefront/core';
+import {AgnosticCoupon, useCartFactory, UseCartFactoryParams, Context} from '@vue-storefront/core';
 
 const getBasketItemByProduct = ({ currentCart, product }) => {
   return currentCart.lineItems.find((item) => item.productId === product._id);
@@ -17,47 +16,47 @@ const getCurrentCart = async (context: Context, currentCart) => {
 };
 
 const params: UseCartFactoryParams<Cart, LineItem, ProductVariant, AgnosticCoupon> = {
-  loadCart: async (context: Context, customQueryFn?: any) => {
+  load: async (context: Context, { customQuery }) => {
     const { $ct } = context;
 
     if ($ct.api.isGuest()) {
       return null;
     }
 
-    const { user } = customQueryFn ? customQueryFn() : { user: null };
+    const { user } = customQuery ? customQuery() : { user: null };
 
     const { data: profileData } = await context.$ct.api.getMe({ customer: false }, user);
 
     return profileData.me.activeCart;
   },
-  addToCart: async (context: Context, { currentCart, product, quantity }, customQuery?: CustomQueryFn) => {
+  addItem: async (context: Context, { currentCart, product, quantity, customQuery }) => {
     const loadedCart = await getCurrentCart(context, currentCart);
 
     const { data } = await context.$ct.api.addToCart(loadedCart, product, quantity, customQuery);
     return data.cart;
   },
-  removeFromCart: async (context: Context, { currentCart, product }, customQuery?: CustomQueryFn) => {
+  removeItem: async (context: Context, { currentCart, product, customQuery }) => {
     const loadedCart = await getCurrentCart(context, currentCart);
 
     const { data } = await context.$ct.api.removeFromCart(loadedCart, product, customQuery);
     return data.cart;
   },
-  updateQuantity: async (context: Context, { currentCart, product, quantity }, customQuery?: CustomQueryFn) => {
+  updateItemQty: async (context: Context, { currentCart, product, quantity, customQuery }) => {
     const loadedCart = await getCurrentCart(context, currentCart);
 
     const { data } = await context.$ct.api.updateCartQuantity(loadedCart, { ...product, quantity }, customQuery);
     return data.cart;
   },
-  clearCart: async (context: Context, { currentCart }) => {
+  clear: async (context: Context, { currentCart }) => {
     return currentCart;
   },
-  applyCoupon: async (context: Context, { currentCart, couponCode }, customQuery?: CustomQueryFn) => {
+  applyCoupon: async (context: Context, { currentCart, couponCode, customQuery }) => {
     const loadedCart = await getCurrentCart(context, currentCart);
 
     const { data } = await context.$ct.api.applyCartCoupon(loadedCart, couponCode, customQuery);
     return { updatedCart: data.cart, updatedCoupon: couponCode };
   },
-  removeCoupon: async (context: Context, { currentCart, coupon }, customQuery?: CustomQueryFn) => {
+  removeCoupon: async (context: Context, { currentCart, coupon, customQuery }) => {
     const loadedCart = await getCurrentCart(context, currentCart);
 
     const { data } = await context.$ct.api.removeCartCoupon(loadedCart, { id: coupon.id, typeId: 'discount-code' }, customQuery);
@@ -68,6 +67,4 @@ const params: UseCartFactoryParams<Cart, LineItem, ProductVariant, AgnosticCoupo
   }
 };
 
-const { useCart } = useCartFactory<Cart, LineItem, ProductVariant, AgnosticCoupon>(params);
-
-export { useCart };
+export default useCartFactory<Cart, LineItem, ProductVariant, AgnosticCoupon>(params);