+++ packages/commercetools/composables/src/useUser/index.ts
@@ -5,6 +5,4 @@ import {
 import { params } from './factoryParams';
 import { useUserFactory } from '@vue-storefront/core';
 
-const { useUser } = useUserFactory<Customer, any, any>(params);
-
-export { useUser };
+export default useUserFactory<Customer, any, any>(params);