+++ packages/commercetools/composables/src/useUserOrders/index.ts
@@ -1,13 +1,12 @@
-import { CustomQuery, useUserOrdersFactory, UseUserOrdersFactoryParams, Context } from '@vue-storefront/core';
+import { useUserOrdersFactory, UseUserOrdersFactoryParams, Context } from '@vue-storefront/core';
 import { Order } from '../types/GraphQL';
 import { OrderSearchParams } from '../types';
 
 const params: UseUserOrdersFactoryParams<Order[], OrderSearchParams> = {
-  searchOrders: async (context: Context, params: OrderSearchParams = {}, customQuery?: CustomQuery): Promise<Order[]> => {
-    const result = await context.$ct.api.getOrders(params, customQuery);
-    const { results } = result.data?.me.orders || { results: [], total: 0 };
-
-    return results;
+  searchOrders: async (context: Context, { customQuery, ...searchParams } = {}): Promise<Order[]> => {
+    const result = await context.$ct.api.getOrders(searchParams, customQuery);
+    const { results: data } = result.data?.me.orders || { results: [], total: 0 };
+    return data;
   }
 };