+++ packages/core/core/src/factories/useProductFactory.ts
@@ -1,25 +1,24 @@
 import { CustomQuery, ProductsSearchParams, UseProduct, Context, FactoryParams } from '../types';
 import { Ref, computed } from '@vue/composition-api';
 import { sharedRef, Logger, generateContext } from '../utils';
-
 export interface UseProductFactoryParams<PRODUCTS, PRODUCT_SEARCH_PARAMS extends ProductsSearchParams> extends FactoryParams {
-  productsSearch: (context: Context, searchParams: PRODUCT_SEARCH_PARAMS, customQuery?: CustomQuery) => Promise<PRODUCTS>;
+  productsSearch: (context: Context, params: PRODUCT_SEARCH_PARAMS & { customQuery?: CustomQuery }) => Promise<PRODUCTS>;
 }
 
 export function useProductFactory<PRODUCTS, PRODUCT_SEARCH_PARAMS>(
   factoryParams: UseProductFactoryParams<PRODUCTS, PRODUCT_SEARCH_PARAMS>
 ) {
-  return function useProduct(id: string): UseProduct<PRODUCTS> {
+  return function useProduct(id: string): UseProduct<PRODUCTS, PRODUCT_SEARCH_PARAMS> {
     const products: Ref<PRODUCTS> = sharedRef([], `useProduct-products-${id}`);
     const loading = sharedRef(false, `useProduct-loading-${id}`);
     const context = generateContext(factoryParams);
 
-    const search = async (params: PRODUCT_SEARCH_PARAMS, customQuery?: CustomQuery) => {
-      Logger.debug('useProduct.search', params);
+    const search = async (searchParams) => {
+      Logger.debug('useProduct.search', searchParams);
 
       loading.value = true;
       try {
-        products.value = await factoryParams.productsSearch(context, params, customQuery);
+        products.value = await factoryParams.productsSearch(context, searchParams);
       } catch (e) {
         Logger.error('useProduct.search', e);
         throw e;